Fixes for previous Q&A and Add supplier relationships for headquarters, categories, and phones
Implements endpoints and service/repository logic to manage many-to-many relationships between suppliers and headquarters, categories, and phones. Includes CRUD operations for these associations, input validation, and handling of inactive or non-existent related entities. Bugs from previous Q&A were fixed, The same logic for create now is implemented for Upgrade. When a taxID is not specified it's value automatically is "Indefinido"